billet edit and delete ok

+++ src/main/java/com/ballo/dao/Dao.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public  interface Dao<T> {
 	//Rechercher
-    public  T recherche(int id);
+    public  T recherche(Long id);
     //inserer
     public  void enregistrer(T obj);
     //modifier
-    public  void modifier(T obj, int id);
+    public  void modifier(T obj, Long id);
     //supprimer
     public  void supprimner(T obj);
     //liste
